Fixing the server unresponsiveness and making the IP dynamic in the mobile app.
The backend had a static locations for the models and the directory did
change so we changed the directories again to the new correct ones.
The server is yet my laptop so the IP is changing from time to time. The
problem is that the IP server in the mobile app is static so an editbox
was added so the user can enter manually the IP address of the server
(The laptop)